Trait midi::ToRawMessages

pub trait ToRawMessages {
    fn to_raw_messages(&self) -> Vec<RawMessage>;
}

Convert self to Vec<RawMessage>

A Vec<RawMessage> represents ordered Midi data that must be sent as a contigious block, this is useful for representing Message::SysEx and Message::NRPN14, note that midi clock messages are allowed to interrupt sysex messages as part of the spec.

Required Methods

fn to_raw_messages(&self) -> Vec<RawMessage>